requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.15.0
httpx==0.27.0
aiosignal==1.3.1
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-SocketIO==5.3.6
frozenlist==1.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
setuptools==69.2.0
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
